normal order

+++ core/domain/payment/payment.go
@@ -80,7 +80,7 @@ func (p *paymentOrderImpl) notifyPaymentFinish() {
 
 	// 通知订单支付完成
 	if p.value.OrderId > 0 {
-		err := p.orderManager.ReceiveNotifyOfOnlineTrade(p.value.OrderId)
+		err := p.orderManager.ReceiveNotifyOfOnlineTrade(int64(p.value.OrderId))
 		domain.HandleError(err, "domain")
 	}
 }
